% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_dep}
\alias{base_dep}
\title{base_dep}
\format{
A data frame with 3624994 rows and 6 variables:
\describe{
\item{\code{sexe}}{int sexe}
\item{\code{prenom}}{chr prenom}
\item{\code{annais}}{factor annee de naissance}
\item{\code{dpt}}{factor departement de naissance}
\item{\code{nombre}}{int nombre de naissances avec ce prenom}
\item{\code{origine}}{ chr origine du prenom}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
base_dep
}
\description{
base departementale Insee des prenoms
}
\keyword{datasets}
