% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_base_dep.R
\name{creation_base_dep}
\alias{creation_base_dep}
\title{Creation de la base departementale à partir de la base mondiale des prenoms
et du fichier departemental insee des prenoms}
\usage{
creation_base_dep(fichier_insee, fichier_prenoms)
}
\arguments{
\item{fichier_insee}{fichier departemental insee des prenoms}

\item{fichier_prenoms}{base mondiale des prenoms}
}
\value{

}
\description{
Creation de la base departementale à partir de la base mondiale des prenoms
et du fichier departemental insee des prenoms
}
