% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_base_nat.R
\name{creation_base_nat}
\alias{creation_base_nat}
\title{Creation de la base nationale à partir de la base mondiale des prenoms
et du fichier national insee des prenoms}
\usage{
creation_base_nat(fichier_insee, fichier_prenoms)
}
\arguments{
\item{fichier_insee}{fichier national insee des prenoms}

\item{fichier_prenoms}{base mondiale des prenoms}
}
\value{
base nationale
}
\description{
Creation de la base nationale à partir de la base mondiale des prenoms
et du fichier national insee des prenoms
}
