% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dpt2018_fichier_initial}
\alias{dpt2018_fichier_initial}
\title{dpt2018_fichier_initial}
\format{
A data frame with 3624994 rows and 5 variables:
\describe{
\item{\code{X.U.FEFF.sexe}}{int sexe}
\item{\code{preusuel}}{factor prenom}
\item{\code{annais}}{factor annee de naissance}
\item{\code{dpt}}{factor departement}
\item{\code{nombre}}{nombre de naissances avec ce prenom}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
dpt2018_fichier_initial
}
\description{
Fichier des prenoms 2018 de l'Insee
}
\keyword{datasets}
