% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_dep}
\alias{map_dep}
\title{map_dep}
\format{
A data frame with 101 rows and 3 variables:
\describe{
\item{\code{codgeo}}{factor code geographique des departements}
\item{\code{libgeo}}{factor libelle des departements}
\item{\code{geometry}}{list geometrie des departements}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
map_dep
}
\description{
carte des departements français
}
\keyword{datasets}
