% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mise_forme_fichier_insee.R
\name{mise_forme_fichier_insee}
\alias{mise_forme_fichier_insee}
\title{mise en forme du fichier insee national ou departemental des prenoms}
\usage{
mise_forme_fichier_insee(fichier_insee)
}
\arguments{
\item{fichier_insee}{fichier national ou departemental insee des prenoms}
}
\value{
dataframe mis en forme pour jointure avec base mondiale des prenoms
}
\description{
mise en forme du fichier insee national ou departemental des prenoms
}
