% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mise_forme_fichier_prenoms.R
\name{mise_forme_fichier_prenoms}
\alias{mise_forme_fichier_prenoms}
\title{mise en forme du fichier des prenoms}
\usage{
mise_forme_fichier_prenoms(fichier_prenoms)
}
\arguments{
\item{fichier_prenoms}{base de données mondiale provenant du site behind the name}
}
\value{
dataframe mis en forme pour jointure avec fichier insee des prenoms
}
\description{
mise en forme du fichier des prenoms
}
