% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat2018_fichier_initial}
\alias{nat2018_fichier_initial}
\title{nat2018_fichier_initial}
\format{
A data frame with 636474 rows and 4 variables:
\describe{
\item{\code{sexe}}{int sexe}
\item{\code{preusuel}}{factor prenom}
\item{\code{annais}}{factor annee de naissance}
\item{\code{nombre}}{nombre de naissances avec ce prenom}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
nat2018_fichier_initial
}
\description{
Fichier des prenoms 2018 de l'Insee
}
\keyword{datasets}
