% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recherche_prenoms_proches.R
\name{recherche_prenoms_proches}
\alias{recherche_prenoms_proches}
\title{Recherche des prenoms proches, selon la distance de Jaro Winkler, dans une base de données}
\usage{
recherche_prenoms_proches(prenom, seuil_graphe, prenoms_all)
}
\arguments{
\item{prenom}{prenom dont on cherche des prenoms proches}

\item{seuil_graphe}{seuil de 0 à 1 en desosus duquel les prenoms sont définis comme proches}

\item{prenoms_all}{liste des prenoms où on cherche des similarités}
}
\value{
une liste avec un vecteur des prenoms proches, un vecteur des distances au prenom d'interet,
et le nb de prénoms proches
}
\description{
Recherche des prenoms proches, selon la distance de Jaro Winkler, dans une base de données
}
