% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prenomfrance.R
\name{run_prenomfrance}
\alias{run_prenomfrance}
\title{run_prenomfrance
Fonction qui permet de lancer l'app shiny}
\usage{
run_prenomfrance()
}
\value{
lance l'app shiny
}
\description{
run_prenomfrance
Fonction qui permet de lancer l'app shiny
}
