% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_nat}
\alias{table_nat}
\title{table_nat}
\format{
A data frame with 119 rows and 44 variables:
\describe{
\item{\code{annee}}{factor annee}
\item{\code{spanish}}{factor proportion de prenoms espagnol }
\item{\code{african}}{factor proportion de prenoms african }
\item{\code{arabic}}{factor proportion de prenoms arabic }
\item{\code{basque}}{factor proportion de prenoms basque }
\item{\code{bosnian}}{factor proportion de prenoms bosnian }
\item{\code{breton}}{factor proportion de prenoms breton }
\item{\code{bulgarian}}{factor proportion de prenoms bulgarian }
\item{\code{catalan}}{factor proportion de prenoms catalan }
\item{\code{celtic}}{factor proportion de prenoms celtic }
\item{\code{croatian}}{factor proportion de prenoms croatian }
\item{\code{czech}}{factor proportion de prenoms czech }
\item{\code{danish}}{factor proportion de prenoms danish }
\item{\code{dutch}}{factor proportion de prenoms dutch }
\item{\code{english}}{factor proportion de prenoms english }
\item{\code{finnish}}{factor proportion de prenoms finnish }
\item{\code{french}}{factor proportion de prenoms french }
\item{\code{german}}{factor proportion de prenoms german }
\item{\code{greek}}{factor proportion de prenoms greek }
\item{\code{hebrew}}{factor proportion de prenoms hebrew }
\item{\code{hungarian}}{factor proportion de prenoms hungarian }
\item{\code{irish}}{factor proportion de prenoms irish }
\item{\code{italian}}{factor proportion de prenoms italian }
\item{\code{icelandic}}{factor proportion de prenoms icelandic }
\item{\code{indian}}{factor proportion de prenoms indian }
\item{\code{japanese}}{factor proportion de prenoms japanese }
\item{\code{jewish}}{factor proportion de prenoms jewish }
\item{\code{macedonian}}{factor proportion de prenoms macedonian }
\item{\code{norwegian}}{factor proportion de prenoms norwegian }
\item{\code{persian}}{factor proportion de prenoms persian }
\item{\code{polish}}{factor proportion de prenoms polish }
\item{\code{portuguese}}{factor proportion de prenoms portuguese }
\item{\code{provençal}}{factor proportion de prenoms provençal }
\item{\code{romanian}}{factor proportion de prenoms romanian }
\item{\code{russian}}{factor proportion de prenoms russian }
\item{\code{scandinavian}}{factor proportion de prenoms scandinavian }
\item{\code{scottish}}{factor proportion de prenoms scottish }
\item{\code{serbian}}{factor proportion de prenoms serbian }
\item{\code{slovak}}{factor proportion de prenoms slovak }
\item{\code{slovene}}{factor proportion de prenoms slovene }
\item{\code{swedish}}{factor proportion de prenoms swedish }
\item{\code{turkish}}{factor proportion de prenoms turkish }
\item{\code{ukrainian}}{factor proportion de prenoms ukrainian }
\item{\code{welsh}}{factor proportion de prenoms welsh }
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
table_nat
}
\description{
table departementale des proportions deprenoms par origine
}
\keyword{datasets}
