% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prenoms_fichier_initial}
\alias{prenoms_fichier_initial}
\title{prenoms_fichier_initial}
\format{
A data frame with 11627 rows and 4 variables:
\describe{
\item{\code{prenom}}{factor prenom}
\item{\code{genre}}{factor sexe}
\item{\code{langage}}{factor origine}
\item{\code{frequence}}{num frequence par million}
}
}
\source{
\url{https://www.data.gouv.fr/fr/datasets/liste-de-prenoms/}
}
\usage{
prenoms_fichier_initial
}
\description{
Auteurs: Mike Campbell et son merveilleux site "Behind The Name"
}
\keyword{datasets}
