% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_dep}
\alias{table_dep}
\title{table_dep}
\format{
A data frame with 11781 rows and 43 variables:
\describe{
\item{\code{annee}}{factor annee}
\item{\code{dep}}{factor departement}
\item{\code{spanish}}{factor proportion de prenoms espagnol par departement}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
table_dep
}
\description{
table departementale des proportions deprenoms par origine
}
\keyword{datasets}
