% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_nat}
\alias{table_nat}
\title{table_nat}
\format{
A data frame with 119 rows and 44 variables:
\describe{
\item{\code{annee}}{factor annee}
\item{\code{spanish}}{factor proportion de prenoms espagnol en France}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
table_nat
}
\description{
table departementale des proportions deprenoms par origine
}
\keyword{datasets}
