% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud2a.R
\name{wordcloud2a}
\alias{wordcloud2a}
\title{Version Wordcloud2a qui permet l'affichage simultané d'un autre plot sur la même page shiny}
\usage{
wordcloud2a(
  data,
  size = 1,
  minSize = 0,
  gridSize = 0,
  fontFamily = "Segoe UI",
  fontWeight = "bold",
  color = "random-dark",
  backgroundColor = "white",
  minRotation = -pi/4,
  maxRotation = pi/4,
  shuffle = TRUE,
  rotateRatio = 0.4,
  shape = "circle",
  ellipticity = 0.65,
  widgetsize = NULL,
  figPath = NULL,
  hoverFunction = NULL
)
}
\arguments{
\item{data}{data}

\item{size}{size}

\item{minSize}{minSize}

\item{gridSize}{gridSize}

\item{fontFamily}{fontFamily}

\item{fontWeight}{fontWeight}

\item{color}{color}

\item{backgroundColor}{backgroundColor}

\item{minRotation}{minRotation}

\item{maxRotation}{maxRotation}

\item{shuffle}{shuffle}

\item{rotateRatio}{rotateRatio}

\item{shape}{shape}

\item{ellipticity}{ellipticity}

\item{widgetsize}{widgetsize}

\item{figPath}{figPath}

\item{hoverFunction}{hoverFunction}
}
\value{
wordcloud2a
}
\description{
Version Wordcloud2a qui permet l'affichage simultané d'un autre plot sur la même page shiny
}
